Added abstract button base class. Added pull up/down support
