+++ devices/Button/Button/Sample/Program.cs
@@ -2,31 +2,31 @@ using Iot.Device.Button;
 using System.Diagnostics;
 using System.Threading;
 
-using Button button = new Button();
+GpioButton button = new GpioButton();
 
 Debug.WriteLine("Button is initialized, starting to read state");
 
 button.IsDoubleClickEnabled = true;
 button.IsHoldingEnabled = true;
 
-button.ButtonDown += (sender, e) =>
-{
-    Debug.WriteLine($"buttondown IsPressed={button.IsPressed}");
-};
+//button.ButtonDown += (sender, e) =>
+//{
+//    Debug.WriteLine($"buttondown IsPressed={button.IsPressed}");
+//};
 
-button.ButtonUp += (sender, e) =>
-{
-    Debug.WriteLine($"buttonup IsPressed={button.IsPressed}");
-};
+//button.ButtonUp += (sender, e) =>
+//{
+//    Debug.WriteLine($"buttonup IsPressed={button.IsPressed}");
+//};
 
 button.Click += (sender, e) =>
 {
-    Debug.WriteLine($"click IsPressed={button.IsPressed}");
+    Debug.WriteLine($"click");
 };
 
 button.DoubleClick += (sender, e) =>
 {
-    Debug.WriteLine($"double click IsPressed={button.IsPressed}");
+    Debug.WriteLine($"double click");
 };
 
 button.Holding += (sender, e) =>
@@ -34,10 +34,10 @@ button.Holding += (sender, e) =>
     switch (e.HoldingState)
     {
         case ButtonHoldingState.Started:
-            Debug.WriteLine($"Holding Started IsPressed={button.IsPressed}");
+            Debug.WriteLine($"Holding Started");
             break;
         case ButtonHoldingState.Completed:
-            Debug.WriteLine($"Holding Completed IsPressed={button.IsPressed}");
+            Debug.WriteLine($"Holding Completed");
             break;
     }
 };
